refactor(highlight): extract incremental and render logic to new files
- Move treeSitterRenderCache and incremental parsing logic to treesitter_incremental.go
- Move render methods (render, renderLines) and highlight query logic to treesitter_render.go
- Remove unused imports (bytes, sort) from main treesitter.go
- Split ~600 lines of code into focused modules for better maintainability
- Keep capture-to-class conversion and lexer resolution in original file
- Add copyright headers to new 2026 files per project guidelines
